Separated session code from traversal code
